Prevent calling waitall() inside a GreenPool's greenthread
When neutron-server is running with several api workers
sending it a termination signal (SIGTERM, SIGHUP or SIGINT)
leads to waitall() being called inside a GreenPool's greenthread.

The reason is that a wsgi server is started in a green thread
from the same green pool that is passed to the server itself
to be used for spawning client green threads.

To avoid it, it is reasonable to use different pools for spawning
a wsgi server and for its internal usage.

This is also the case for metadata agent running with several
metadata workers.

Upstream-Kilo: https://review.openstack.org/#/c/157320
Resolves rhbz: 1298598
Closes-Bug: #1423250
(cherry picked from commit f9b2791735863b58171ee8be46c47ef3ed2c5edd)

Change-Id: I38174396f06fcb29ac0776534ac6494dabb00df6
Reviewed-on: https://code.engineering.redhat.com/gerrit/65635
Tested-by: Arie Bregman <abregman@redhat.com>
Tested-by: RHOS Jenkins <apevec+rhosci@redhat.com>
Reviewed-by: Ihar Hrachyshka <ihrachys@redhat.com>