Set loading strategy to joined for Routerport/Port
The RouterPort model has a relationship to the ports model which
is frequently relied on to get the port IDs of interfaces attached
to a router. However, this defaults to the loading strategy to
'select', which meant a new query was being emitted for every
interface to the ports table just to get the ID.

This patch adjusts the relationship to be 'joined' by default so
one query will fetch the related ports.

Another option would have been not to use the port object at all since
the ID is all that the callers were usually interested in. However,
they would end up using the ID to do a port lookup, which is being
optimized away in another patch anyway so the full port object from
the relationship will end up getting used.

Upstream-Juno: https://review.openstack.org/#/c/175589/
Related rhbz: 1281583
Change-Id: Id1ae35f845f7367d5f1f065c6fa637da7b980a2b
Partial-Bug: #1445412
(cherry picked from commit 10b17a884452736a6b214bcb7705b955192a1748)
Reviewed-on: https://code.engineering.redhat.com/gerrit/61614
Reviewed-by: Miguel Angel Ajo Pelayo <majopela@redhat.com>
Tested-by: RHOS Jenkins <apevec+rhosci@redhat.com>