Change setup.cfg to match upstream tarball
Because of setup.cfg contains TABs instead of spaces as separator in
upstream tarball, patches touching setup.cfg are not compatible. This
patch makes setup.cfg compatible as per [1].

This patch makes the tabs inside of setup.cfg - so after this patch,
we're the same as tarball. But later revert is needed.

[1] https://www.rdoproject.org/packaging/rdo-packaging.html#_modifying_the_package

Change-Id: I01b19c35a292606aed21c8faaa394fd55df2fdd3