Don't return exception when attempting to add duplicate VIP
In commit 8d71521f30277937d1951be2ba90cdaf4a2a12ed we throw
an exception when trying to add a keepalived VIP that is
already present in the agent's internal cache. Since this
operation should be idempotent like other L3 agent operations,
don't raise an exception, but write a DEBUG log instead.

OSP 6: The aforementioned commit was downstream only backported,
and this patch is needed.
OSP 7: Commit 8d7 was not backported and so duplicate VIPs
would be added to keepalived.conf, which will be fixed.
Master: Being handled in
https://review.openstack.org/#/c/254995/ which will be
backported to OSP 8.

Downstream-Only
Closes rhbz: 1275020
Change-Id: Ib1fb1a2e46c326075450921d604ad73d399b9cf3
Reviewed-on: https://code.engineering.redhat.com/gerrit/63338
Reviewed-by: Jakub Libosvar <jlibosva@redhat.com>
Tested-by: Jakub Libosvar <jlibosva@redhat.com>