ML2: update port's status to DOWN if its binding info has changed
This fixes the problem that when two or more ports in a network
are migrated to a host that did not previously have any ports in
the same network, the new host is sometimes not told about the
IP/MAC addresses of all the other ports in the network.
In other words, initial L2population does not work, for the new host.

This is because the l2pop mechanism driver only sends catch-up
information to the host when it thinks it is dealing with the first
active port on that host; and currently, when multiple ports are
migrated to a new host, there is always more than one active port
so the condition above is never triggered.

The fix is for the ML2 plugin to set a port's status to DOWN when
its binding info changes. This patch also fixes the bug when nova
thinks it should not wait for any events from neutron because all
ports are already active.

Resolves rhbz: 1289995
Upstream-Mitaka: https://review.openstack.org/#/c/215467
Change-Id: I0b190c34f73e42283fe72d09ecc5078a4ef46764